program/requirements.txt
mysql-connector-python==8.0.29
pandas
pyyaml
